Created libCUserNotification
Moved the print_help function back, updated makefile to build a static library that is then included in notifier.c
